fix(android): correct BluetoothGatt APIs for API 34 + FrameProtocol + QRScanner imports
- BleServer: onClientConnectionStateChange -> onConnectionStateChange
- onCharacteristicWriteRequest: replace valueOffset/properties with preparedWrite/responseNeeded
- onCharacteristicReadRequest: swap characteristic/offset parameter order
- onDescriptorWriteRequest: replace valueOffset/properties with preparedWrite/responseNeeded
- Remove setDeviceName from AdvertiseData.Builder (removed in API 34)
- notifyCharacteristicChanged: add device parameter + iterate connectedDevices
- FrameProtocol: add .toByte() for MSG_MODE_SWITCH in byteArrayOf
- QRScannerActivity: add missing Intent import

+++ mobile-app/app/src/main/java/com/ppink/mobile/FrameProtocol.kt
@@ -33,7 +33,7 @@ object FrameProtocol {
 
     /** Creates a mode switch frame */
     fun createModeSwitch(mode: Int): ByteArray {
-        return byteArrayOf(MSG_MODE_SWITCH, mode.toByte())
+        return byteArrayOf(MSG_MODE_SWITCH.toByte(), mode.toByte())
     }
 
     /** Parses an incoming frame from PC (stroke rendering data) */
